add permission check

+++ models/file.go
@@ -1,7 +1,7 @@
 package models
 
 type Metadata struct {
-	Path string
+	Path    string
 	OwnerID string
-	Size int32
+	Size    float64
 }
